products done

+++ lib/models/product.dart
@@ -24,7 +24,7 @@ class Product {
     return Product(
       id: json["id"], 
       title: json["title"], 
-      price: json["price"] , 
+      price: json["price"].toDouble(), 
       description: json["description"] , 
       category: json["category"] , 
       image: json["image"] , 
@@ -46,7 +46,7 @@ class Rating{
 
    factory Rating.fromJson(Map<String, dynamic> json){
      return Rating(
-       rate: json["rate"], 
+       rate: json["rate"].toDouble(), 
        count: json["count"]);
    } 
 }
